3.0.0 Breaking changes to R4 Capability - Renames to `CapabilityStatement` - Fixes `@Pattern` constraints incorrectly applied to `List<String>`
