+++ fhir-common/pom.xml
@@ -8,6 +8,6 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>fhir-common</artifactId>
-  <version>2.0.38-SNAPSHOT</version>
+  <version>3.0.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 </project>
