+++ r4-examples/pom.xml
@@ -8,7 +8,7 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>r4-examples</artifactId>
-  <version>2.0.38-SNAPSHOT</version>
+  <version>3.0.0-SNAPSHOT</version>
   <packaging>jar</packaging>
   <dependencies>
     <dependency>
