--- r4/src/test/java/gov/va/api/health/r4/api/CapabilityTest.java
@@ -1,12 +1,0 @@
-package gov.va.api.health.r4.api;
-
-import gov.va.api.health.r4.api.samples.SampleCapabilities;
-import org.junit.Test;
-
-public class CapabilityTest {
-
-  @Test
-  public void capability() {
-    RoundTrip.assertRoundTrip(SampleCapabilities.get().capability());
-  }
-}
