+++ r4/src/test/java/gov/va/api/health/r4/api/resources/CapabilityStatementTest.java
@@ -1,0 +1,13 @@
+package gov.va.api.health.r4.api.resources;
+
+import gov.va.api.health.r4.api.RoundTrip;
+import gov.va.api.health.r4.api.samples.SampleCapabilityStatements;
+import org.junit.Test;
+
+public class CapabilityStatementTest {
+
+  @Test
+  public void capability() {
+    RoundTrip.assertRoundTrip(SampleCapabilityStatements.get().capability());
+  }
+}
