+++ r4/src/test/java/gov/va/api/health/r4/api/samples/SampleTerminologyCapabilities.java
@@ -2,7 +2,7 @@ package gov.va.api.health.r4.api.samples;
 
 import static java.util.Collections.singletonList;
 
-import gov.va.api.health.r4.api.resources.Capability;
+import gov.va.api.health.r4.api.resources.CapabilityStatement;
 import gov.va.api.health.r4.api.resources.TerminologyCapabilities;
 import java.util.Arrays;
 import java.util.List;
@@ -43,7 +43,7 @@ public class SampleTerminologyCapabilities {
         .version("1")
         .name("conformin' norman")
         .title("urgent care capability")
-        .status(Capability.Status.active)
+        .status(CapabilityStatement.Status.active)
         .experimental("true")
         .date("2000-01-01T00:00:00-00:00")
         .publisher("random house")
@@ -53,7 +53,7 @@ public class SampleTerminologyCapabilities {
         .jurisdiction(singletonList(codeableConcept()))
         .purpose("words words")
         .copyright("Alphonso, Lord of the Mangos")
-        .kind(Capability.Kind.capability)
+        .kind(CapabilityStatement.Kind.capability)
         .software(software())
         .implementation(implementation())
         .lockedDate("true")
