+++ us-core-r4-examples/pom.xml
@@ -8,7 +8,7 @@
     <relativePath/>
   </parent>
   <artifactId>us-core-r4-examples</artifactId>
-  <version>2.0.38-SNAPSHOT</version>
+  <version>3.0.0-SNAPSHOT</version>
   <packaging>jar</packaging>
   <properties/>
   <dependencies>
