+++ us-core-r4/pom.xml
@@ -8,7 +8,7 @@
     <relativePath/>
   </parent>
   <artifactId>us-core-r4</artifactId>
-  <version>2.0.38-SNAPSHOT</version>
+  <version>3.0.0-SNAPSHOT</version>
   <packaging>jar</packaging>
   <properties/>
   <dependencies>
